chore: use fixtures!
